+++ Firmware/BoardSupport/src/disp_t16x7b8x4.c
@@ -66,8 +66,8 @@
                        DISP_LED_MAX_SEL)
 
 /* Display Maximum Value Per Display declaraion */
-#define UPPER_MAX_VAL (9999999)
-#define LOWER_MAX_VAL (9999)
+#define UPPER_MAX_VAL (9999999uL)
+#define LOWER_MAX_VAL (9999uL)
 
 /* Private function prototypes -----------------------------------------------*/
 static void DispInit(void);
@@ -128,8 +128,8 @@ const DISP_TYPE_T DispInfoT16x7B8x4 =
   .uppMaxSel = UPPER_MAX_SEL,
   .lowSegType = DISP_SEVEN_SEG,
   .lowMaxSel = LOWER_MAX_SEL,
-	.uppMaxVal = UPPER_MAX_VAL,
-	.lowMaxVal = LOWER_MAX_VAL,
+	.uppMaxVal = (uint32_t)UPPER_MAX_VAL,
+	.lowMaxVal = (uint32_t)LOWER_MAX_VAL,
   .pLedBitmap = &LED_BITMAP[0],
   .uppFmtStr = "%7lu",
   .lowFmtStr = "%4lu",
